pdf-converter-api/requirements.txt
PyPDF2==3.0.1
python-docx==1.1.2
reportlab==4.0.9
Pillow==10.1.0
